+++ enzyme/c_template/stage.h
@@ -10,18 +10,18 @@ inline void ${STAGE_NAME}(uint64_t NI, uint64_t NJ, uint64_t NK, Workspace * p)
 
     workspace_swap_sync(p, NUM_INPUTS);
 
-    float * p_source = p->source_workspace;
-    float * p_sink = p->sink_workspace;
+    double * p_source = p->source_workspace;
+    double * p_sink = p->sink_workspace;
 
     FOR_IJK {
-        const float * source =    p_source + OFFSET(i,  j,k,NUM_INPUTS);
-        const float * source_ip = p_source + OFFSET(i+1,j,k,NUM_INPUTS);
-        const float * source_im = p_source + OFFSET(i-1,j,k,NUM_INPUTS);
-        const float * source_jp = p_source + OFFSET(i,j+1,k,NUM_INPUTS);
-        const float * source_jm = p_source + OFFSET(i,j-1,k,NUM_INPUTS);
-        const float * source_kp = p_source + OFFSET(i,j,k+1,NUM_INPUTS);
-        const float * source_km = p_source + OFFSET(i,j,k-1,NUM_INPUTS);
-        float * sink = p_sink + OFFSET(i,j,k,NUM_OUTPUTS);
+        const double * source =    p_source + OFFSET(i,  j,k,NUM_INPUTS);
+        const double * source_ip = p_source + OFFSET(i+1,j,k,NUM_INPUTS);
+        const double * source_im = p_source + OFFSET(i-1,j,k,NUM_INPUTS);
+        const double * source_jp = p_source + OFFSET(i,j+1,k,NUM_INPUTS);
+        const double * source_jm = p_source + OFFSET(i,j-1,k,NUM_INPUTS);
+        const double * source_kp = p_source + OFFSET(i,j,k+1,NUM_INPUTS);
+        const double * source_km = p_source + OFFSET(i,j,k-1,NUM_INPUTS);
+        double * sink = p_sink + OFFSET(i,j,k,NUM_OUTPUTS);
         ${CODE}
     }
 }
